third_party/WebKit/LayoutTests/fast/dom/Element/innerHTML-treat-null-as-string.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(() => {
  let p = document.createElement("p");
  p.innerHTML = "null";
  assert_equals(p.innerText, "null");
}, "Setting innerHTML to the string null doesn't result in an empty element");
</script>
